Move algorithm config to ftensp config files
