+++ src/fractool/cli_parser.cpp
@@ -1,5 +1,6 @@
 // Fractool
 #include <fractool/colormap.hpp>
+#include <fractool/ftensp/config.hpp>
 #include <fractool/fractool/cli_parser.hpp>
 
 // External
@@ -40,35 +41,6 @@ static void print_cmaps() {
     }
     std::cout << std::endl;
 }
-
-/**
- * Print algorithm list
- */
-static void print_algos() {
-    BOOST_LOG_TRIVIAL(debug) << "Printing algorithms list";
-    std::cout << std::endl 
-        << "Algorithms list:" 
-        << std::endl 
-        << std::endl;
-    for (auto it = algo_lookup.begin(); it != algo_lookup.end(); ++it) {
-        std::cout << "\t- " << it->first << std::endl;
-    }
-    std::cout << std::endl;
-}
-
-/**
- * Operator to interpret algorithm from input string
- */
-std::istream& operator>> (std::istream &in, ALGORITHM &algorithm)
-{
-    algorithm = ALGORITHM::_INVALID;
-    std::string str;
-    in >> str;
-    auto it = algo_lookup.find(str);
-    if (it != algo_lookup.end()) { algorithm = it->second; }
-    else { BOOST_LOG_TRIVIAL(error) << "Invalid algorithm type: " << str; }
-    return in;
-};
 
 /**
  * Operator to inrerpret colormap from input string
